# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backlogg API",
    "description": "Backlogg is a unified catalog API for **movies, series, books and games**,\nseeded from public external APIs (TMDB, Open Library, IGDB) via a nightly sync,\nplus a social layer on top of that catalog.\n\n**Catalog & discovery:** list and detail endpoints per content type, cross-type\n`/search` (with on-demand external fallback), `/genres`, `/trending`, `/people`\nand similar-item recommendations.\n\n**Social layer:** account management with refresh tokens and account recovery\n(`/auth`, `/users`), ratings & reviews, per-user library/backlog, curated\n`/lists`, `/follows`, an activity `/feed`, `/notifications` and personalized\n`/recommendations`.\n\n**Platform:** rate limiting, structured observability, Prometheus `/metrics`\nand response caching.\n\n## Authentication\n\n- **User endpoints** use `Authorization: Bearer <access_token>`. Obtain the\n  token pair from `POST /auth/login`; renew it with `POST /auth/refresh`.\n- **Admin endpoints** (`/admin/*`) require the `X-API-Key` header.\n\nEverything else is public. Direct messaging between users is out of scope.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/games": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List games",
        "description": "Paginated list of catalogued games (no external fallback); genre filter + sort.",
        "operationId": "list_games_v1_games_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by genre slug",
              "title": "Genre"
            },
            "description": "Filter by genre slug"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GameSortEnum",
              "description": "Sort order",
              "default": "rating_desc"
            },
            "description": "Sort order"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/games/{slug}/ratings": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List game ratings",
        "description": "Public, paginated ratings & reviews for a game, newest first.",
        "operationId": "list_game_ratings_v1_games__slug__ratings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/games/{slug}/rating": {
      "put": {
        "tags": [
          "games"
        ],
        "summary": "Rate a game",
        "description": "Upsert the caller's score/review (full replace); recomputes aggregates. Auth.",
        "operationId": "rate_game_v1_games__slug__rating_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Delete own game rating",
        "description": "Remove the caller's rating and recompute aggregates. Requires auth.",
        "operationId": "delete_game_rating_v1_games__slug__rating_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/games/{slug}/library": {
      "put": {
        "tags": [
          "games"
        ],
        "summary": "Set game library status",
        "description": "Upsert the caller's backlog status for this game. Requires auth.",
        "operationId": "set_game_library_v1_games__slug__library_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryEntryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Remove game from library",
        "description": "Delete the caller's library entry for this game. Requires auth.",
        "operationId": "delete_game_library_v1_games__slug__library_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/games/{slug}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Get game detail",
        "description": "Full game detail; on-demand IGDB fallback. Auth optional (adds viewer_status).",
        "operationId": "get_game_v1_games__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameGenreOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "GameGenreOut"
      },
      "GameListItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "poster_url",
          "release_date",
          "rating_external",
          "genres"
        ],
        "title": "GameListItemOut"
      },
      "GameListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameListItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "GameListOut"
      },
      "GameOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "game_type": {
            "type": "string",
            "title": "Game Type"
          },
          "original_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Language"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "rating_count_external": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Count External"
          },
          "rating_internal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Internal"
          },
          "rating_count_internal": {
            "type": "integer",
            "title": "Rating Count Internal"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/GameGenreOut"
            },
            "type": "array",
            "title": "Genres"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/GamePlatformOut"
            },
            "type": "array",
            "title": "Platforms"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/backlogg__shared__schemas__CreditOut"
            },
            "type": "array",
            "title": "Credits",
            "default": []
          },
          "viewer_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "original_title",
          "slug",
          "overview",
          "release_date",
          "game_type",
          "original_language",
          "poster_url",
          "backdrop_url",
          "rating_external",
          "rating_count_external",
          "rating_internal",
          "rating_count_internal",
          "genres",
          "platforms"
        ],
        "title": "GameOut"
      },
      "GamePlatformOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "GamePlatformOut"
      },
      "GameSortEnum": {
        "type": "string",
        "enum": [
          "rating_desc",
          "rating_asc",
          "date_desc",
          "date_asc",
          "title_asc"
        ],
        "title": "GameSortEnum"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryEntryIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LibraryStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LibraryEntryIn",
        "description": "Request body for PUT /{type}/{slug}/library."
      },
...
      "LibraryStatus": {
        "type": "string",
        "enum": [
          "want",
          "in_progress",
          "completed",
          "dropped"
        ],
        "title": "LibraryStatus",
        "description": "Backlog status values accepted in request bodies and filters."
      },
      "LibraryStatusOut": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "slug",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LibraryStatusOut",
        "description": "Response of PUT /{type}/{slug}/library \u2014 the caller's own entry.",
        "example": {
          "created_at": "2026-05-25T18:04:11Z",
          "item_type": "MOVIE",
          "slug": "dune-2021",
          "status": "want",
          "updated_at": "2026-05-25T18:04:11Z"
        }
      },
...
      "RatingAuthorOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "avatar_url"
        ],
        "title": "RatingAuthorOut"
      },
      "RatingIn": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          }
        },
        "type": "object",
        "title": "RatingIn",
        "description": "Request body for PUT /{type}/{slug}/rating \u2014 full replace (upsert)."
      },
      "RatingListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RatingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "RatingListOut"
      },
      "RatingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/RatingAuthorOut"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "liked_by_viewer": {
            "type": "boolean",
            "title": "Liked By Viewer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "score",
          "review_text",
          "like_count",
          "liked_by_viewer",
          "created_at",
          "updated_at"
        ],
        "title": "RatingOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "backlogg__shared__schemas__CreditOut": {
        "properties": {
          "person_name": {
            "type": "string",
            "title": "Person Name"
          },
          "person_slug": {
            "type": "string",
            "title": "Person Slug"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "character_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Name"
          },
          "billing_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Order"
          }
        },
        "type": "object",
        "required": [
          "person_name",
          "person_slug",
          "profile_url",
          "role",
          "character_name",
          "billing_order"
        ],
        "title": "CreditOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}